refactor: split api/routes.py into domain-scoped routers
Split the 1,872-line monolithic routes.py into 9 focused domain routers
under api/routes/ and introduce the services/orchestration/ package.

Domain routers:
- pets_crud.py   — CRUD, feed, my-pets, my-repos
- pets_appearance.py — style, name, badge-style, skins
- pets_media.py  — badge SVG, images, animated GIFs, generate/regenerate
- pets_info.py   — characteristics, comments, achievements, milestones,
                   contributors, blame-board
- pets_actions.py — resurrect
- admin.py       — pet admin settings, contributor exclusion, reset/delete
- social.py      — leaderboard, showcase SVG, contributor badge SVG
- system.py      — styles, badge-styles, image-provider health, queue stats
- webhooks.py    — GitHub webhook receiver

Also adds:
- api/dependencies.py: shared get_pet_or_404 helper
- services/orchestration/__init__.py: placeholder for future extraction
- specs/001-api-layered-architecture/: spec, plan, tasks documentation

All 35 endpoint URLs preserved. 1121/1121 tests pass. Lint clean.

Patchable symbols (settings, StorageService, etc.) are re-exported from
the package __init__ so existing test patches require no modification.
