+++ .specify/feature.json
@@ -1,0 +1,3 @@
+{
+  "feature_directory": "specs/001-api-layered-architecture"
+}
