+++ specs/001-api-layered-architecture/plan.md
@@ -1,0 +1,154 @@
+# Implementation Plan: API Layered Architecture
+
+**Branch**: `001-api-layered-architecture` | **Date**: 2026-04-13 | **Spec**: [spec.md](./spec.md)
+
+## Summary
+
+Split `api/routes.py` (1,872 lines, 35 endpoints) into domain-scoped routers and introduce an orchestration service layer. HTTP handlers become thin pass-throughs; business logic moves into injectable service classes that delegate to existing `services/` and `crud/` layers. Zero behaviour change — all URLs, request shapes, and response shapes are preserved.
+
+## Technical Context
+
+**Language/Version**: Python 3.11
+**Primary Dependencies**: FastAPI, SQLAlchemy async, Pydantic v2
+**Storage**: PostgreSQL (via existing `crud/` layer and `AsyncSession`)
+**Testing**: pytest + pytest-asyncio, existing `TestClient` fixtures
+**Target Platform**: Linux (k3s)
+**Project Type**: Web service (API layer refactor only)
+**Performance Goals**: No change — refactor must not degrade response times
+**Constraints**: All 35 endpoint URLs unchanged; all 1,100+ tests pass without modification
+**Scale/Scope**: 35 endpoints → 6 domain routers + 7 orchestration services
+
+## Constitution Check
+
+| Principle | Status | Notes |
+|-----------|--------|-------|
+| Repo-health-as-pet-health | ✅ Pass | Refactor only — no behaviour change |
+| Deterministic identity | ✅ Pass | No changes to appearance/personality logic |
+| Forgiving mechanics | ✅ Pass | No changes to health/death logic |
+| Badge-first externalisability | ✅ Pass | Badge endpoints preserved at same URLs |
+| Observability | ✅ Pass | Logging/metrics hooks preserved |
+| Minimal blast radius | ✅ Pass | Incremental domain-by-domain split |
+| Definition of Done | ✅ Pass | All tests must pass before merge |
+
+No constitution violations.
+
+## Project Structure
+
+### Documentation (this feature)
+
+```
+specs/001-api-layered-architecture/
+├── plan.md              ← this file
+├── research.md          ← domain boundaries and dependency graph
+├── data-model.md        ← service interface contracts
+└── tasks.md             ← generated by /speckit-tasks
+```
+
+### Source Code — after refactor
+
+```
+src/github_tamagotchi/
+├── api/
+│   ├── __init__.py
+│   ├── auth.py               (unchanged)
+│   ├── health.py             (unchanged)
+│   ├── alerts.py             (unchanged)
+│   ├── routes/               (NEW — replaces routes.py)
+│   │   ├── __init__.py       (registers all sub-routers)
+│   │   ├── pets_crud.py      (POST/GET/DELETE /pets, GET /me/pets)
+│   │   ├── pets_appearance.py (style, name, badge-style, skins)
+│   │   ├── pets_media.py     (images, GIFs, badges, generate/regen)
+│   │   ├── pets_info.py      (characteristics, comments, achievements, milestones, contributors, blame-board)
+│   │   ├── pets_actions.py   (feed, resurrect)
+│   │   ├── admin.py          (admin settings, contributor exclusion, reset, delete)
+│   │   ├── social.py         (leaderboard, showcase, contributor badge SVG)
+│   │   ├── system.py         (styles, badge-styles, health/image-provider, queue/stats)
+│   │   └── webhooks.py       (POST /webhooks/github)
+│   └── dependencies.py       (NEW — shared Depends helpers)
+│
+├── services/
+│   ├── (all existing files — UNCHANGED)
+│   └── orchestration/        (NEW — thin orchestration layer)
+│       ├── __init__.py
+│       ├── pet_image_service.py
+│       ├── animated_gif_service.py
+│       ├── pet_resurrection_service.py
+│       ├── webhook_processor_service.py
+│       ├── contributor_badge_service.py
+│       ├── leaderboard_service.py
+│       └── pet_admin_service.py
+│
+├── crud/                     (UNCHANGED)
+└── models/                   (UNCHANGED)
+
+tests/
+├── (all existing tests — UNCHANGED, imports updated if needed)
+└── unit/services/orchestration/  (NEW unit tests per service)
+```
+
+## Phase 0: Domain Boundaries
+
+### Router → endpoint mapping
+
+| Router | Endpoints |
+|--------|-----------|
+| `pets_crud.py` | POST /pets, GET /pets, GET /pets/{o}/{r}, DELETE /pets/{o}/{r}, GET /me/pets, POST /pets/{o}/{r}/feed |
+| `pets_appearance.py` | PUT style, PUT name, PUT badge-style, GET skins, PUT skin |
+| `pets_media.py` | GET badge.svg, GET image/{stage}, GET image/{stage}/animated, POST generate-images, POST regenerate-images |
+| `pets_info.py` | GET characteristics, GET+POST comments, GET achievements, GET milestones, GET contributors, GET blame-board |
+| `pets_actions.py` | POST resurrect |
+| `admin.py` | GET/PATCH/DELETE /admin, POST/DELETE contributors/exclude, POST reset |
+| `social.py` | GET leaderboard, GET showcase/{u}.svg, GET contributor/{o}/{r}/{u}.svg |
+| `system.py` | GET styles, GET badge-styles, GET health/image-provider, GET admin/queue/stats |
+| `webhooks.py` | POST /webhooks/github |
+
+### Orchestration service → extracted logic
+
+| Service | Lines extracted from routes.py | Responsibility |
+|---------|-------------------------------|----------------|
+| `PetImageService` | 1007-1080 (~73 lines) | Storage lookup → on-demand generation → fallback |
+| `AnimatedGifService` | 1082-1201 (~119 lines) | Sprite sheet gen → GIF composition → caching |
+| `PetResurrectionService` | 475-536 (~61 lines) | Grace period check → state reset → image re-queue |
+| `WebhookProcessorService` | 1346-1437 (~91 lines) | Signature verify → event dispatch → payload summary |
+| `ContributorBadgeService` | 1526-1593 (~67 lines) | Contributor stats → standing determination → badge SVG |
+| `LeaderboardService` | 1456-1484 (~28 lines) | Category definitions → ranked queries |
+| `PetAdminService` | 1596-1872 (~276 lines) | Settings CRUD, contributor exclusion, reset |
+
+### Shared dependency helpers in `api/dependencies.py`
+
+- `get_pet_or_404(repo_owner, repo_name, session)` — fetch pet or raise 404
+- `require_repo_admin(pet, current_user)` — raise 403 if not owner/admin (consolidates duplicated check)
+- `get_optional_current_user` — already exists in auth.py, re-export here
+
+## Phase 1: Service Interface Contracts
+
+```python
+class PetImageService:
+    async def get_or_generate(repo_owner, repo_name, stage, session, storage) -> bytes | None
+
+class AnimatedGifService:
+    async def get_or_generate(repo_owner, repo_name, stage, pet, session, storage) -> bytes
+
+class PetResurrectionService:
+    async def resurrect(pet, session, now) -> Pet  # raises ResurrectionNotAllowed
+
+class WebhookProcessorService:
+    async def process(payload, signature, event_type, session) -> dict  # raises InvalidSignature
+
+class ContributorBadgeService:
+    async def get_badge_svg(repo_owner, repo_name, username, github_service) -> bytes
+
+class LeaderboardService:
+    async def get_leaderboard(session) -> list[dict]
+
+class PetAdminService:
+    async def get_settings(pet, session) -> dict
+    async def update_settings(pet, data, session) -> dict
+    async def exclude_contributor(pet, github_login, session) -> None
+    async def remove_exclusion(pet, github_login, session) -> None
+    async def reset_pet(pet, session) -> Pet
+```
+
+## Complexity Tracking
+
+No constitution violations — no entries needed.
