+++ src/github_tamagotchi/services/orchestration/__init__.py
@@ -1,0 +1,5 @@
+"""Orchestration service layer.
+
+These services coordinate domain logic between existing services, CRUD layers, and storage.
+They are called by HTTP handlers; handlers remain thin pass-throughs.
+"""
